- Añadidas lectura de parámetros de configuración para gráfica1 y gráfica2 	- Añadida visualización de media total el gráfica2 	- Corregido bug al liberar imagen.

+++  trackingdrosophila --username chaos.ruben@gmail.com/Tracking/Current/Visualizacion.hpp
@@ -98,7 +98,7 @@ typedef struct graphBarsParams{
 
 	CvPoint Origen;
 	CvPoint fin;
-	int escalaY;
+	float escalaY;
 	int puntosY;
 	int puntosX;
 	int maxYval;
@@ -123,7 +123,7 @@ typedef struct graphMovParams{
 	int periodoFr;
 	CvPoint Origen;
 	CvPoint fin;
-	int escalaY;
+	float escalaY;
 	int puntosY;
 	int puntosX;
 	int maxYval;
